requirements.txt
altgraph==0.17.2
amqp==5.1.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==4.1.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
croniter==1.3.5
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.10.1
django-drf-filepond==0.4.1
django-filter==21.1
django-storages==1.13.1
djangorestframework==3.12.4
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.15.3
Flask==2.2.2
fonttools==4.31.2
future==0.18.2
gevent==21.12.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.4.2
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.4
mljar-mercury==1.1.5
mysql-connector-python==8.0.31
nbclient==0.5.13
nbconvert==7.0.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.21.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pdfminer==20191125
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.8.0
pycparser==2.21
pycryptodome==3.14.1
pyee==8.2.2
Pygments==2.11.2
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
pyocr==0.8.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyppeteer==1.0.2
pyrsistent==0.18.1
PySimpleGUI==4.60.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.1.5
Send2Trash==1.8.0
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
SQLAlchemy==1.4.27
sqlparse==0.4.2
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
whitenoise==5.3.0
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
